iotcloud-device/requirements.txt
paho-mqtt
falcon
gunicorn
-e libs/samsungctl
-e libs/flux_led